requirements.txt
scapy
python-nmap
python-nmap==0.7.1
